feat: 추천 score 테이블 관련
