+++ backend/src/main/java/com/meong9/backend/domain/recommendation/id_class/PensionPlaceId.java
@@ -10,6 +10,7 @@ import java.io.Serializable;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EqualsAndHashCode
 @Getter
+@ToString
 public class PensionPlaceId implements Serializable {
     private Long pensionId;
     private Long placeId;
