+++ backend/src/main/java/com/meong9/backend/domain/recommendation/member_score/pension_place_score/entity/PensionPlaceScore.java
@@ -14,6 +14,7 @@ import java.time.LocalDateTime;
 @EqualsAndHashCode
 @Getter
 @Builder
+@ToString
 public class PensionPlaceScore {
     @EmbeddedId
     private PensionPlaceId pensionPlaceId;
